+++ src/Graph/Nodes/Node.php
@@ -2,7 +2,15 @@
 
 namespace Rubix\ML\Graph\Nodes;
 
+use Generator;
+
 interface Node
 {
-    //
+    /**
+     * Return a generator for all of the node's edges i.e. the nodes that
+     * this node connects to.
+     *
+     * @return Generator
+     */
+    public function edges() : Generator;
 }